requirements.txt
youtube-transcript-api
python-dotenv
marvin
typer
